pyproject.toml
[tool.poetry]
package-mode=false

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.15.2"
django-model-utils = "^5.0.0"
drf-spectacular = "^0.28.0"

[tool.poetry.group.dev.dependencies]
pylint = "^3.3.1"
black = "^24.10.0"
isort = "^5.13.2"
pylint-django = "^2.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# black
[tool.black]
line-length = 79
extend-exclude = "migrations/"

# isort
[tool.isort]
profile = "black"
known_local_folder=["api"]
line_length = 79
lines_after_imports = 2
extend_skip_glob = "migrations"


[tool.pylint]
ignore = ["migrations","manage.py", "settings.py", ".git", "__pycache__", "migrations", ".venv"]
max-line-length = 79
disable = ["C0114", "C0115", "C0116", "W1514", "W3101", "R0903", "C0103"]